Keep file length budget: split missing-bundle tests into their own suite
workflow-guard-tests flagged the previous commit: GhosttyTerminalView.swift
went 4 lines over its checked-in length budget and
ShellStartupMatrixTests.swift crossed the 500-line tracking threshold.

Fold the sidebarShellIntegration read into the spawn-env guard so the call
site change is net-zero lines in GhosttyTerminalView.swift, and move the
four deleted-bundle degradation tests into a new wired
ShellStartupMissingBundleTests suite (own private helpers, matching the
existing per-file runProcess pattern) so the matrix file stays under the
threshold. No behavior change.
